app/backend/backend_shared/pyproject.toml
[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[project]
name = "backend-shared"
version = "0.1.0"
description = "Shared utilities for our FastAPI services (Clean Architecture/SOLID)."
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Team" }]
dependencies = [
  "pydantic>=2.5",
  "typing-extensions>=4.10.0",
  "fastapi>=0.104.0",
  "starlette>=0.27.0",
  "pyyaml>=6.0",
  "pandas>=2.0.0",
  "SQLAlchemy>=2.0.25",
  "asyncpg>=0.29.0",
  "google-cloud-storage>=2.10.0",
]

[project.optional-dependencies]
dev = ["pytest", "build", "twine", "python-dotenv"]

[tool.hatch.build]
exclude = [
  "/tests",
  "*/__pycache__/*",
]

[tool.hatch.build.targets.wheel]
packages = ["src/backend_shared"]
